fix(dashboard): 修复数据统计和查询逻辑并增强安全性
- 修复热门需求统计SQL返回类型不匹配问题，将COUNT结果转换为SIGNED类型
- 修复热力图数据查询逻辑，统一使用逻辑删除标记并优化日期参数处理
- 修复网络关系数据查询，移除冗余的关联条件并确保数据类型一致性
- 为所有仪表板接口添加身份验证要求，启用方法级安全配置
- 优化服务层实现，统一逻辑删除时间常量并改进错误处理

+++ zhilian-backend/src/main/java/com/zhilian/zhilianbackend/dto/response/TopDemandResponse.java
@@ -25,5 +25,5 @@ public class TopDemandResponse {
     /**
      * 需求数量
      */
-    private Long count;
+    private Integer count;
 }
